ADD : DishPage

+++ src/main/java/com/cloudstone/emenu/storage/db/util/SqlUtils.java
@@ -20,23 +20,23 @@ public class SqlUtils {
         return flag? 1 : 0;
     }
     
-    public static int[] strToIds(String str) {
+    public static long[] strToIds(String str) {
         if (StringUtils.isBlank(str)) {
-            return new int[0];
+            return new long[0];
         } else {
             String[] ss = str.split(",");
-            int[] r = new int[ss.length];
+            long[] r = new long[ss.length];
             for (int i=0; i<r.length; i++) {
-                r[i] = Integer.valueOf(ss[i]);
+                r[i] = Long.valueOf(ss[i]);
             }
             return r;
         }
     }
     
-    public static String idsToStr(int[] ids) {
+    public static String idsToStr(long[] ids) {
         StringBuilder sb = new StringBuilder();
         boolean first = true;
-        for (int id:ids) {
+        for (long id:ids) {
             if (!first) {
                 sb.append(",");
             }
